Рефакторинг, добавление смарт компонента
